+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to the "unsaved-files-vscode" extension will be documented i
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## 2.1.2 - 2020-12-??
+## 2.1.2 - 2021-10-??
+
+### Added
+
+- context menu in view on explorer.
 
 ### Changed